+++ Source/TurboSequence_Lf/Public/TurboSequence_Manager_Lf.h
@@ -128,7 +128,7 @@ public:
 
 protected:
 	// NODE: Please use the versions with FTurboSequence_MeshSpawnData_Lf& FromSpawnData
-	static uint32 AddSkinnedMeshInstance_GameThread(const TObjectPtr<UTurboSequence_MeshAsset_Lf> FromAsset,
+	static int32 AddSkinnedMeshInstance_GameThread(const TObjectPtr<UTurboSequence_MeshAsset_Lf> FromAsset,
 	                                                const FTransform& SpawnTransform,
 	                                                const TObjectPtr<UWorld> InWorld,
 	                                                const TArray<TObjectPtr<UMaterialInterface>>& OverrideMaterials =
@@ -136,7 +136,7 @@ protected:
 	                                                const TObjectPtr<UTurboSequence_FootprintAsset_Lf> FootprintAsset =
 		                                                nullptr);
 
-	static bool RemoveSkinnedMeshInstance_GameThread(int64 MeshID, const TObjectPtr<UWorld> InWorld);
+	static bool RemoveSkinnedMeshInstance_GameThread(int32 MeshID, const TObjectPtr<UWorld> InWorld);
 
 public:
 	/**
@@ -155,7 +155,7 @@ protected:
 public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Add, Update, Mesh, Group"))
-	static void AddInstanceToUpdateGroup_RawID_Concurrent(const int32 GroupIndex, int64 MeshID);
+	static void AddInstanceToUpdateGroup_RawID_Concurrent(const int32 GroupIndex, int32 MeshID);
 
 	/**
 	 * Adds an Mesh to the Update Group with the given Index
@@ -169,7 +169,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Remove, Update, Mesh, Group"))
-	static void RemoveInstanceFromUpdateGroup_RawID_Concurrent(const int32 GroupIndex, int64 MeshID);
+	static void RemoveInstanceFromUpdateGroup_RawID_Concurrent(const int32 GroupIndex, int32 MeshID);
 
 
 	/**
@@ -185,7 +185,7 @@ public:
 	/**
 	 * Get the Number of Mesh Collections in a Update Group with the given Index, useful if you need iterate
 	 * @param GroupIndex The Group Index
-	 * @return The Number of Mesh Collections in a Update Group
+	 * @return The Number of Mesh Collections in an Update Group
 	 */
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Number, Num, Update, Mesh, Group"))
@@ -194,7 +194,7 @@ public:
 	/**
 	 * Get the Number of Mesh IDs in a Update Group with the given Index, useful if you need iterate
 	 * @param GroupIndex The Group Index
-	 * @return The Number of Mesh IDs in a Update Group
+	 * @return The Number of Mesh IDs in an Update Group
 	 */
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Number, Num, Update, Mesh, Group"))
@@ -237,7 +237,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="World Space Transform", Keywords="Turbo, Sequence, TS, Get, Transform, World, Mesh"))
-	static FTransform GetMeshWorldSpaceTransform_RawID_Concurrent(int64 MeshID)
+	static FTransform GetMeshWorldSpaceTransform_RawID_Concurrent(int32 MeshID)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -265,7 +265,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Camera Distance", Keywords="Turbo, Sequence, TS, Get, Distance, Camera, Mesh"))
-	static float GetMeshClosestCameraDistance_RawID_Concurrent(int64 MeshID)
+	static float GetMeshClosestCameraDistance_RawID_Concurrent(int32 MeshID)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -291,7 +291,7 @@ public:
 		{
 			SetMeshWorldSpaceLocationRotationScale_RawID_Concurrent(MeshData.RootMotionMeshID, Location, Rotation,
 			                                                        Scale);
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				SetMeshWorldSpaceLocationRotationScale_RawID_Concurrent(MeshID, Location, Rotation, Scale);
 			}
@@ -300,7 +300,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Set, Animation, Mesh, Motion, Transform"))
-	static void SetMeshWorldSpaceLocationRotationScale_RawID_Concurrent(int64 MeshID, const FVector Location,
+	static void SetMeshWorldSpaceLocationRotationScale_RawID_Concurrent(int32 MeshID, const FVector Location,
 	                                                                    const FQuat Rotation, const FVector Scale)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
@@ -332,7 +332,7 @@ public:
 		if (MeshData.IsMeshDataValid())
 		{
 			SetMeshWorldSpaceTransform_RawID_Concurrent(MeshData.RootMotionMeshID, Transform);
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				SetMeshWorldSpaceTransform_RawID_Concurrent(MeshID, Transform);
 			}
@@ -341,7 +341,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Set, Animation, Mesh, Motion, Transform"))
-	static void SetMeshWorldSpaceTransform_RawID_Concurrent(int64 MeshID, const FTransform Transform)
+	static void SetMeshWorldSpaceTransform_RawID_Concurrent(int32 MeshID, const FTransform Transform)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -382,7 +382,7 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Set, Animation, Root, Motion, Transform"))
 	static bool GetRootMotionTransform_RawID_Concurrent(FTransform& OutRootMotion,
-	                                                    int64 MeshID,
+	                                                    int32 MeshID,
 	                                                    float DeltaTime,
 	                                                    const EBoneSpaces::Type& Space = EBoneSpaces::WorldSpace)
 	{
@@ -410,7 +410,7 @@ public:
 	 * @param MeshData The Mesh ID
 	 * @param DeltaTime The Delta Time which the Root Motion needs to get extracted, Use your Group Delta Time or the Frame Delta Time
 	 * @param bZeroZAxis Do you want Zero Z values in the Component Space Root Motion Transform
-	 * @param bIncludeScale Do you want include the Root Motion Scale3D Factor which will scale the mesh
+	 * @param bIncludeScale Do you want to include the Root Motion Scale3D Factor which will scale the mesh
 	 */
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Set, Animation, Root, Motion, Transform"))
@@ -423,7 +423,7 @@ public:
 		{
 			MoveMeshWithRootMotion_RawID_Concurrent(MeshData.RootMotionMeshID, DeltaTime, bZeroZAxis, bIncludeScale);
 			const FTransform& MeshTransform = GetMeshWorldSpaceTransform_RawID_Concurrent(MeshData.RootMotionMeshID);
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				SetMeshWorldSpaceTransform_RawID_Concurrent(MeshID, MeshTransform);
 			}
@@ -432,7 +432,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(Keywords="Turbo, Sequence, TS, Set, Animation, Root, Motion, Transform"))
-	static void MoveMeshWithRootMotion_RawID_Concurrent(int64 MeshID,
+	static void MoveMeshWithRootMotion_RawID_Concurrent(int32 MeshID,
 	                                                    float DeltaTime,
 	                                                    bool ZeroZAxis = false,
 	                                                    bool bIncludeScale = false)
@@ -468,7 +468,7 @@ public:
 
 	/**
 	 * Get the Number of all meshes no matter which update group
-	 * @return The Number of all all instance meshes, no matter which Update Group
+	 * @return The Number of all instance meshes, no matter which Update Group
 	 */
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Num Meshes", Keywords="Turbo, Sequence, TS, Get, Number, Instances"))
@@ -496,7 +496,7 @@ public:
 		{
 			FTurboSequence_AnimMinimalCollection_Lf Data = FTurboSequence_AnimMinimalCollection_Lf(true);
 			Data.RootMotionMesh = PlayAnimation_RawID_Concurrent(MeshData.RootMotionMeshID, Animation, AnimSettings);
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				Data.CustomizableMeshes.Add(PlayAnimation_RawID_Concurrent(MeshID, Animation, AnimSettings));
 			}
@@ -507,7 +507,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Animation Data", Keywords="Turbo, Sequence, TS, Set, Animation, PLay"))
-	static FTurboSequence_AnimMinimalData_Lf PlayAnimation_RawID_Concurrent(int64 MeshID,
+	static FTurboSequence_AnimMinimalData_Lf PlayAnimation_RawID_Concurrent(int32 MeshID,
 	                                                                        UAnimSequence* Animation,
 	                                                                        const FTurboSequence_AnimPlaySettings_Lf&
 	                                                                        AnimSettings
@@ -556,7 +556,7 @@ public:
 			FTurboSequence_AnimMinimalBlendSpaceCollection_Lf Data =
 				FTurboSequence_AnimMinimalBlendSpaceCollection_Lf(true);
 			Data.RootMotionMesh = PlayBlendSpace_RawID_Concurrent(MeshData.RootMotionMeshID, BlendSpace, AnimSettings);
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				Data.CustomizableMeshes.Add(PlayBlendSpace_RawID_Concurrent(MeshID, BlendSpace, AnimSettings));
 			}
@@ -568,7 +568,7 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Blend Space Data", Keywords=
 			"Turbo, Sequence, TS, Set, Animation, PLay, Blend Space, Blend"))
-	static FTurboSequence_AnimMinimalBlendSpace_Lf PlayBlendSpace_RawID_Concurrent(int64 MeshID,
+	static FTurboSequence_AnimMinimalBlendSpace_Lf PlayBlendSpace_RawID_Concurrent(int32 MeshID,
 		UBlendSpace* BlendSpace,
 		const FTurboSequence_AnimPlaySettings_Lf& AnimSettings
 	)
@@ -611,7 +611,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Animation Sequence", Keywords="Turbo, Sequence, TS, Get, Animation, Play, Priority"))
-	static UAnimSequence* GetHighestPriorityPlayingAnimation_RawID_Concurrent(int64 MeshID)
+	static UAnimSequence* GetHighestPriorityPlayingAnimation_RawID_Concurrent(int32 MeshID)
 	{
 		if (!GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -673,7 +673,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Set, Animation, Settings, Tweaks, Blend"))
-	static bool TweakBlendSpace_RawID_Concurrent(int64 MeshID,
+	static bool TweakBlendSpace_RawID_Concurrent(int32 MeshID,
 	                                             const FTurboSequence_AnimMinimalBlendSpace_Lf& BlendSpaceData,
 	                                             const FVector3f& WantedPosition)
 	{
@@ -803,7 +803,7 @@ public:
 				return true;
 			}
 
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				if (GetIKTransform_RawID_Concurrent(OutIKTransform, MeshID, BoneName, AnimationDeltaTime, Space))
 				{
@@ -817,7 +817,7 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Get, Transform, IK, Bone"))
 	static bool GetIKTransform_RawID_Concurrent(FTransform& OutIKTransform,
-	                                            int64 MeshID,
+	                                            int32 MeshID,
 	                                            const FName& BoneName,
 	                                            float AnimationDeltaTime,
 	                                            const EBoneSpaces::Type& Space = EBoneSpaces::WorldSpace)
@@ -871,7 +871,7 @@ public:
 				return true;
 			}
 
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				if (SetIKTransform_RawID_Concurrent(MeshID, BoneName, IKTransform, Space))
 				{
@@ -884,7 +884,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Set, Transform, IK, Bone"))
-	static bool SetIKTransform_RawID_Concurrent(int64 MeshID,
+	static bool SetIKTransform_RawID_Concurrent(int32 MeshID,
 	                                            const FName& BoneName,
 	                                            const FTransform& IKTransform,
 	                                            const EBoneSpaces::Type& Space = EBoneSpaces::WorldSpace)
@@ -938,7 +938,7 @@ public:
 				return true;
 			}
 
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				if (GetSocketTransform_RawID_Concurrent(OutSocketTransform, MeshID, SocketName, AnimationDeltaTime,
 				                                        Space))
@@ -953,7 +953,7 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Get, Socket, Transform, IK, Item"))
 	static bool GetSocketTransform_RawID_Concurrent(FTransform& OutSocketTransform,
-	                                                int64 MeshID,
+	                                                int32 MeshID,
 	                                                const FName& SocketName,
 	                                                float AnimationDeltaTime,
 	                                                const EBoneSpaces::Type& Space = EBoneSpaces::WorldSpace)
@@ -1002,7 +1002,7 @@ public:
 			{
 				return true;
 			}
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				if (GetIsMeshVisibleInCameraFrustum_RawID_Concurrent(MeshID))
 				{
@@ -1015,7 +1015,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Is Visisble", Keywords="Turbo, Sequence, TS, Get, Visible, Camera, Frustum"))
-	static bool GetIsMeshVisibleInCameraFrustum_RawID_Concurrent(int64 MeshID)
+	static bool GetIsMeshVisibleInCameraFrustum_RawID_Concurrent(int32 MeshID)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -1049,7 +1049,7 @@ public:
 			{
 				return true;
 			}
-			for (int64 MeshID : MeshData.CustomizableMeshIDs)
+			for (int32 MeshID : MeshData.CustomizableMeshIDs)
 			{
 				if (GetReferencePoseTransform_RawID_Concurrent(OutRefPoseTransform, MeshID, BoneName,
 				                                               MeshTransformSpace))
@@ -1064,7 +1064,7 @@ public:
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Get, Animation, Reference, Pose, Transform"))
 	static bool GetReferencePoseTransform_RawID_Concurrent(FTransform& OutRefPoseTransform,
-	                                                       int64 MeshID,
+	                                                       int32 MeshID,
 	                                                       const FName& BoneName,
 	                                                       const ETurboSequence_TransformSpace_Lf& MeshTransformSpace =
 		                                                       ETurboSequence_TransformSpace_Lf::WorldSpace)
@@ -1125,7 +1125,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Mesh Asset", Keywords="Turbo, Sequence, TS, Get, Mesh, Asset"))
-	static UTurboSequence_MeshAsset_Lf* GetMeshAsset_RawID_Concurrent(int64 MeshID)
+	static UTurboSequence_MeshAsset_Lf* GetMeshAsset_RawID_Concurrent(int32 MeshID)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -1136,7 +1136,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Set, Customize, Change, Mesh, Material"))
-	static bool CustomizeMesh_RawID_GameThread(int64 MeshID,
+	static bool CustomizeMesh_RawID_GameThread(int32 MeshID,
 	                                           UTurboSequence_MeshAsset_Lf* TargetMesh,
 	                                           const TArray<UMaterialInterface*>& TargetMaterials)
 	{
@@ -1226,7 +1226,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Curve", Keywords="Turbo, Sequence, TS, Get, Animation, Curve"))
-	static FTurboSequence_PoseCurveData_Lf GetAnimationCurveValue_RawID_Concurrent(int64 MeshID,
+	static FTurboSequence_PoseCurveData_Lf GetAnimationCurveValue_RawID_Concurrent(int32 MeshID,
 		const FName& CurveName,
 		UAnimSequence* Animation)
 	{
@@ -1288,7 +1288,7 @@ public:
 
 	UFUNCTION(BlueprintPure, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Footprint Asset", Keywords="Turbo, Sequence, TS, Get, Footpint, Asset"))
-	static UTurboSequence_FootprintAsset_Lf* GetFootprintAsset_RawID_Concurrent(int64 MeshID)
+	static UTurboSequence_FootprintAsset_Lf* GetFootprintAsset_RawID_Concurrent(int32 MeshID)
 	{
 		if (GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
 		{
@@ -1327,7 +1327,7 @@ public:
 
 	UFUNCTION(BlueprintCallable, Category="Turbo Sequence",
 		meta=(ReturnDisplayName="Success", Keywords="Turbo, Sequence, TS, Set, Custom, Data, Rendering, Per Instance"))
-	static bool SetCustomDataToInstance_RawID_Concurrent(int64 MeshID, uint8 CustomDataFractionIndex,
+	static bool SetCustomDataToInstance_RawID_Concurrent(int32 MeshID, uint8 CustomDataFractionIndex,
 	                                                     float CustomDataValue)
 	{
 		if (!GlobalLibrary.RuntimeSkinnedMeshes.Contains(MeshID))
@@ -1361,7 +1361,7 @@ public:
 
 		bool bSuccess = SetCustomDataToInstance_RawID_Concurrent(MeshData.RootMotionMeshID, CustomDataFractionIndex,
 		                                                         CustomDataValue);
-		for (int64 MeshID : MeshData.CustomizableMeshIDs)
+		for (int32 MeshID : MeshData.CustomizableMeshIDs)
 		{
 			SetCustomDataToInstance_RawID_Concurrent(MeshID, CustomDataFractionIndex, CustomDataValue);
 		}
